src/third_party/WebKit/LayoutTests/editing/shadow/contenteditable-propagation-to-distributed-node-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks that contenteditable is propagated to distributed nodes.

PASS getComputedStyle(shadowRoot.querySelector('span'), null).webkitUserModify is "read-only"
PASS getComputedStyle(shadowRoot.querySelector('content').getDistributedNodes()[0], null).webkitUserModify is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
